AiService/requirements.txt
Flask
Flask-Cors
openai
langchain
langchain-community
chromadb
python-dotenv
openai
PyPDF2